feat: add household finance CRUD, CSV import/export, and account deletion
Adds income, expense, budget, and savings-goal endpoints on top of the
ownership-enforced repository base: recurring entries store a rule once and
have their monthly occurrences computed on demand (packages/analytics), so
"how much recurs into this month" is never duplicated as stored rows.
Budgets expose computed planned-vs-actual status (spent, remaining,
over-budget); savings goals compute current amount, progress, and a
projected completion date from their contribution ledger rather than a
single running total.

CSV import is a real preview-then-commit flow: preview validates a file
(missing fields, bad amounts/dates, unknown categories, duplicates against
existing rows) without writing anything; commit re-validates and writes
only the valid, non-duplicate rows, returning a per-row report either way.
CSV export sanitizes cells starting with =, +, -, or @ against formula
injection. JSON account export covers every owned entity; DELETE /api/users/
me lets a user delete their own account (fastapi-users only ships a
superuser-gated delete-by-id route) and cascades all of their data.

Fixes a real route-ordering bug caught by the account-deletion tests: the
custom self-delete route has to be registered before fastapi-users' own
users router, otherwise its superuser-gated DELETE /{id} greedily matches
"me" as a literal id first and returns 403 instead of ever reaching ours.

52 backend tests passing; verified end-to-end against the running Docker
container (demo household's seeded budget shows correct actual/remaining).

+++ apps/api/app/models/finance.py
@@ -16,6 +16,7 @@ from __future__ import annotations
 import enum
 import uuid
 from datetime import date
+from decimal import Decimal
 
 from sqlalchemy import Date, Enum, ForeignKey, Numeric, String, Text
 from sqlalchemy.dialects.postgresql import UUID
@@ -85,7 +86,7 @@ class IncomeEntry(Base, UUIDPrimaryKeyMixin, TimestampMixin):
         UUID(as_uuid=True), ForeignKey("category.id", ondelete="SET NULL"), nullable=True
     )
     label: Mapped[str] = mapped_column(String(200), nullable=False)
-    amount: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False)
+    amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
     entry_date: Mapped[date] = mapped_column(Date, nullable=False)
     is_recurring: Mapped[bool] = mapped_column(default=False, nullable=False)
     recurrence_rule_id: Mapped[uuid.UUID | None] = mapped_column(
@@ -109,7 +110,7 @@ class ExpenseEntry(Base, UUIDPrimaryKeyMixin, TimestampMixin):
         UUID(as_uuid=True), ForeignKey("category.id", ondelete="SET NULL"), nullable=True
     )
     label: Mapped[str] = mapped_column(String(200), nullable=False)
-    amount: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False)
+    amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
     entry_date: Mapped[date] = mapped_column(Date, nullable=False)
     is_recurring: Mapped[bool] = mapped_column(default=False, nullable=False)
     recurrence_rule_id: Mapped[uuid.UUID | None] = mapped_column(
@@ -132,7 +133,7 @@ class Budget(Base, UUIDPrimaryKeyMixin, TimestampMixin):
     category_id: Mapped[uuid.UUID] = mapped_column(
         UUID(as_uuid=True), ForeignKey("category.id", ondelete="CASCADE"), nullable=False
     )
-    monthly_limit: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False)
+    monthly_limit: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
     effective_from: Mapped[date] = mapped_column(Date, nullable=False)
     effective_to: Mapped[date | None] = mapped_column(Date, nullable=True)
 
@@ -144,7 +145,7 @@ class SavingsGoal(Base, UUIDPrimaryKeyMixin, TimestampMixin):
         UUID(as_uuid=True), ForeignKey("user.id", ondelete="CASCADE"), nullable=False
     )
     name: Mapped[str] = mapped_column(String(200), nullable=False)
-    target_amount: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False)
+    target_amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
     target_date: Mapped[date | None] = mapped_column(Date, nullable=True)
     category_id: Mapped[uuid.UUID | None] = mapped_column(
         UUID(as_uuid=True), ForeignKey("category.id", ondelete="SET NULL"), nullable=True
@@ -163,6 +164,6 @@ class SavingsGoalContribution(Base, UUIDPrimaryKeyMixin, TimestampMixin):
     savings_goal_id: Mapped[uuid.UUID] = mapped_column(
         UUID(as_uuid=True), ForeignKey("savings_goal.id", ondelete="CASCADE"), nullable=False
     )
-    amount: Mapped[float] = mapped_column(Numeric(12, 2), nullable=False)
+    amount: Mapped[Decimal] = mapped_column(Numeric(12, 2), nullable=False)
     contributed_on: Mapped[date] = mapped_column(Date, nullable=False)
     note: Mapped[str | None] = mapped_column(Text, nullable=True)
